S302T300-56 #time 45m ?implement ?test Create Inventory Post Controller Endpoint - Added Added method stubs in service interface and impl (for testing) - Added JUnit tests - Added creation DTO and changed ZonedDateTime to LocalDate as no time is given.

+++ backend/src/main/java/com/navbara_pigeons/wasteless/dto/BasicInventoryDto.java
@@ -11,8 +11,8 @@ public class BasicInventoryDto {
   private long id;
   private BasicProductDto product;
   private long quantity;
-  private float pricePerItem;
-  private float totalPrice;
+  private Double pricePerItem;
+  private Double totalPrice;
   private ZonedDateTime manufactured;
   private ZonedDateTime sellBy;
   private ZonedDateTime bestBefore;
